pyproject.toml
[project]
name = "summarizer"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Ynakatsuka", email = "nk.tsssa@gmail.com" }
]
dependencies = [
    "google-generativeai>=0.8.0",
    "requests>=2.32.3",
    "beautifulsoup4>=4.12.3",
    "pillow>=10.4.0",
    "pypdf2>=3.0.1",
    "pandas>=2.2.2",
    "pytest-playwright>=0.5.2",
    "pyyaml>=6.0.2",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/summarizer"]
